requirements.txt
dj-database-url==0.4.2
Django==1.11.1
django-auth-ldap==1.2.12
django-registration==2.1.2
djangorestframework==3.5.1
gunicorn==19.7.0
psycopg2==2.6.2
python-http-client==2.2.1
python-ldap==2.4.39
pytz==2017.2
requests==2.12.3
six==1.10.0
urllib3==1.19.1
whitenoise==3.3.0
